Fix warning
git-svn-id: https://svn.osgeo.org/gdal/trunk@40983 f0d54148-0727-0410-94bb-9a71ac55c965

+++ autotest/cpp/tut/tut_posix.hpp
@@ -344,7 +344,7 @@ private:
 
             //TODO: max failure length
             char buffer[1024];
-            int r = read(pipe, buffer, sizeof(buffer));
+            int r = static_cast<int>(read(pipe, buffer, sizeof(buffer)));
             ensure_errno("sanity check on read() failed", r >= 0);
 
             if(r > 0)
